Factored out multi-size logic into its own module.
